+++ connector/src/test/scala/com/microsoft/kusto/spark/WriterTests.scala
@@ -7,7 +7,7 @@ import java.util
 import java.util.TimeZone
 import java.util.zip.GZIPOutputStream
 
-import com.microsoft.kusto.spark.datasink.{BlobWriteResource, CountingWriter, KustoSinkOptions, KustoWriter}
+import com.microsoft.kusto.spark.datasink._
 import com.microsoft.kusto.spark.datasource.KustoSourceOptions
 import com.microsoft.kusto.spark.utils.{KustoDataSourceUtils => KDSU}
 import org.apache.commons.lang3.time.FastDateFormat
@@ -47,7 +47,7 @@ class WriterTests extends FlatSpec with Matchers {
     val streamWriter = new OutputStreamWriter(byteArrayOutputStream)
     val writer = new BufferedWriter(streamWriter)
     val csvWriter = CountingWriter(writer)
-    KustoWriter.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
     writer.flush()
     writer.close()
     byteArrayOutputStream.toString() shouldEqual "\"John,\"\" Doe\",1" + lineSep
@@ -63,7 +63,7 @@ class WriterTests extends FlatSpec with Matchers {
     val streamWriter = new OutputStreamWriter(byteArrayOutputStream)
     val writer = new BufferedWriter(streamWriter)
     var csvWriter = CountingWriter(writer)
-    KustoWriter.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
     writer.flush()
     val got = byteArrayOutputStream.toString()
 
@@ -80,7 +80,7 @@ class WriterTests extends FlatSpec with Matchers {
     val writer = new BufferedWriter(streamWriter)
     val csvWriter = CountingWriter(writer)
 
-    KustoWriter.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
     writer.flush()
     writer.close()
     csvWriter.getCounter shouldEqual expectedSize
@@ -144,16 +144,16 @@ class WriterTests extends FlatSpec with Matchers {
     // 2018-06-18 16:04:00.0 in GMT
     val ts = new java.sql.Timestamp(1529337840000l)
 
-    val someData2 = List(
+    val someDateData = List(
       Map("asd" -> Row(Date.valueOf("1991-09-07"), ts, false, java.math.BigDecimal.valueOf(1/100000.toDouble)))
     )
 
-    val someData3 = List(
+    val someEmptyArrays = List(
       Row(Row(Array(null,""),"")),
       Row(Row(null,""))
     )
 
-    val someData4 = List(
+    val someDecimalData = List(
       Row(BigDecimal("123456789.123456789")),
       Row(BigDecimal("123456789123456789.123456789123456789")),
       Row(BigDecimal("-123456789123456789.123456789123456789")),
@@ -167,13 +167,13 @@ class WriterTests extends FlatSpec with Matchers {
     val someSchema = List(
       StructField("mapToArray", MapType(StringType, new StructType().add("arrayStrings", ArrayType(StringType, containsNull = true), nullable = true), valueContainsNull = true), nullable = true)
     )
-    val someSchema2 = List(
+    val someDateSchema = List(
       StructField("mapToStruct", MapType(StringType, new StructType().add("date", DateType, nullable = true).add("time", TimestampType).add("booly", BooleanType).add("deci", DataTypes.createDecimalType(20,14)), valueContainsNull = true), nullable = true)
     )
-    val someSchema3 = List(
+    val someEmptyArraysSchema = List(
       StructField("emptyStruct", new StructType().add("emptyArray", ArrayType(StringType, containsNull = true), nullable = true).add("emptyString",StringType))
     )
-    val someSchema4 = List(
+    val someDecimalSchema = List(
       StructField("BigDecimals", DataTypes.createDecimalType(38,10))
     )
 
@@ -182,25 +182,25 @@ class WriterTests extends FlatSpec with Matchers {
       StructType(WriterTests.asSchema(someSchema))
     )
 
-    val df2 = sparkSession.createDataFrame(
-      sparkSession.sparkContext.parallelize(WriterTests.asRows(someData2)),
-      StructType(WriterTests.asSchema(someSchema2))
+    val dateDf = sparkSession.createDataFrame(
+      sparkSession.sparkContext.parallelize(WriterTests.asRows(someDateData)),
+      StructType(WriterTests.asSchema(someDateSchema))
     )
 
-    val df3 = sparkSession.createDataFrame(
-      sparkSession.sparkContext.parallelize(WriterTests.asRows(someData3)),
-      StructType(WriterTests.asSchema(someSchema3))
+    val emptyArraysDf = sparkSession.createDataFrame(
+      sparkSession.sparkContext.parallelize(WriterTests.asRows(someEmptyArrays)),
+      StructType(WriterTests.asSchema(someEmptyArraysSchema))
     )
 
-    val df4 = sparkSession.createDataFrame(
-      sparkSession.sparkContext.parallelize(WriterTests.asRows(someData4)),
-      StructType(WriterTests.asSchema(someSchema4))
+    val someDecimalDf = sparkSession.createDataFrame(
+      sparkSession.sparkContext.parallelize(WriterTests.asRows(someDecimalData)),
+      StructType(WriterTests.asSchema(someDecimalSchema))
     )
 
     val dfRow: InternalRow = df.queryExecution.toRdd.collect().head
-    val dfRow2 = df2.queryExecution.toRdd.collect.head
-    val dfRows3 = df3.queryExecution.toRdd.collect.take(2)
-    val dfRows4 = df4.queryExecution.toRdd.collect.take(someData4.length)
+    val dfRow2 = dateDf.queryExecution.toRdd.collect.head
+    val dfRows3 = emptyArraysDf.queryExecution.toRdd.collect.take(2)
+    val dfRows4 = someDecimalDf.queryExecution.toRdd.collect.take(someDecimalData.length)
 
     val dateFormat = FastDateFormat.getInstance("yyyy-MM-dd'T'HH:mm:ss.SSSXXX", TimeZone.getTimeZone("GMT"))
 
@@ -209,36 +209,36 @@ class WriterTests extends FlatSpec with Matchers {
     val writer = new BufferedWriter(streamWriter)
     val csvWriter = CountingWriter(writer)
 
-    KustoWriter.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRow, df.schema, dateFormat, csvWriter)
 
     writer.flush()
     val res1 = byteArrayOutputStream.toString
     res1 shouldEqual "\"{\"\"asd\"\":{\"\"arrayStrings\"\":[\"\"stringVal\\n\\r\\\\\\\"\"\"\"]},\"\"asd2\"\":{\"\"arrayStrings\"\":[\"\"stringVal2\\b\\f\"\"]}}\"" + lineSep
 
     byteArrayOutputStream.reset()
-    KustoWriter.writeRowAsCSV(dfRow2, df2.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRow2, dateDf.schema, dateFormat, csvWriter)
     writer.flush()
     val res2 = byteArrayOutputStream.toString
     res2 shouldEqual "\"{\"\"asd\"\":{\"\"date\"\":\"\"1991-09-07\"\",\"\"time\"\":\"\"2018-06-18T16:04:00.000Z\"\",\"\"booly\"\":false,\"\"deci\"\":\"\"0.00001000000000\"\"}}\"" + lineSep
 
     byteArrayOutputStream.reset()
-    KustoWriter.writeRowAsCSV(dfRows3(0), df3.schema, dateFormat, csvWriter)
-    KustoWriter.writeRowAsCSV(dfRows3(1), df3.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRows3(0), emptyArraysDf.schema, dateFormat, csvWriter)
+    RowCSVWriterUtils.writeRowAsCSV(dfRows3(1), emptyArraysDf.schema, dateFormat, csvWriter)
     writer.flush()
     val res3 = byteArrayOutputStream.toString
     res3 shouldEqual "\"{\"\"emptyArray\"\":[null,\"\"\"\"],\"\"emptyString\"\":\"\"\"\"}\"" + lineSep + "\"{\"\"emptyString\"\":\"\"\"\"}\"" + lineSep
 
     byteArrayOutputStream.reset()
     for(row<- dfRows4){
-      KustoWriter.writeRowAsCSV(row, df4.schema, dateFormat, csvWriter)
+      RowCSVWriterUtils.writeRowAsCSV(row, someDecimalDf.schema, dateFormat, csvWriter)
     }
 
     writer.flush()
     writer.close()
     val res4 = byteArrayOutputStream.toString
-    res4 shouldEqual "\"123456789123456789.1234567891\"" + lineSep + "\"123456789123456789.1234567891\"" + lineSep +
-      "\"0.1234567891\"" + lineSep + "\"0.1234567891\"" + lineSep + "\"0.0000000000\"" + lineSep + "\"0.0000000000\"" + lineSep +
-      "\"-0.1000000000\"" + lineSep + "\"-0.1000000000\"" + lineSep
+    res4 shouldEqual "\"123456789.1234567890\"" + lineSep + "\"123456789123456789.1234567891\"" + lineSep + "\"-123456789123456789.1234567891\"" + lineSep +
+      "\"0.1234567891\"" + lineSep + "\"-123456789123456789.0000000000\"" + lineSep + "\"0.0000000000\"" + lineSep +
+      "\"0.1000000000\"" + lineSep + "\"-0.1000000000\"" + lineSep
   }
 }
